% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment7.R
\name{segment7}
\alias{segment7}
\title{my_function}
\usage{
segment7(
  value = 8,
  dp = FALSE,
  color = "#00FF00",
  stroke = "#004D00",
  background = "#000000",
  height = 320
)
}
\value{
A character vector
}
\description{
A that returns a seven-segment display as an SVG
}
\examples{
segment7(8)
}
