% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment7.R
\name{segment7}
\alias{segment7}
\title{my_function}
\usage{
segment7(
  value = 8,
  dp = FALSE,
  color = "#00FF00",
  stroke = "#004D00",
  background = "#000000",
  height = 320
)
}
\arguments{
\item{value}{A single digit (0-9) to be displayed. Defaults to 8.}

\item{dp}{A boolean indicating whether to display the decimal point. Defaults to FALSE}

\item{color}{The color of the active segments. Specifieds as a hex code. Defaults to "#00FF00".}

\item{stroke}{The color of the stroke around the segments. Specified as a hex code. Defaults to "#004D00".}

\item{background}{The background color of the SVG. Specified as a hex code. Defaults to "#000000".}

\item{height}{The height of the SVG in pixels. Defaults to 320.}
}
\value{
A character string containing the SVG representation of the seven-segment display.
}
\description{
A that returns a seven-segment display as an SVG
}
\examples{
segment7(value = 3, dp = TRUE,
    color = "#FF0000", stroke = "#000000", background = "#FFFFFF",
    height = 200)
}
